Baseline: https://github.com/paradigmxyz/reth/pull/12046

```console
update rlp node level/size 100000 | depth 0
                        time:   [55.814 ms 59.301 ms 63.485 ms]
                        change: [+50.942% +60.183% +72.344%] (p = 0.00 < 0.05)
                        Performance has regressed.
update rlp node level/size 100000 | depth 1
                        time:   [30.273 ms 30.783 ms 31.239 ms]
                        change: [-21.595% -20.500% -19.435%] (p = 0.00 < 0.05)
                        Performance has improved.
update rlp node level/size 100000 | depth 2
                        time:   [31.512 ms 32.004 ms 32.396 ms]
                        change: [-26.814% -25.234% -23.831%] (p = 0.00 < 0.05)
                        Performance has improved.
update rlp node level/size 100000 | depth 3
                        time:   [32.535 ms 32.858 ms 33.157 ms]
                        change: [-71.004% -70.568% -70.180%] (p = 0.00 < 0.05)
                        Performance has improved.
update rlp node level/size 100000 | depth 4
                        time:   [39.292 ms 39.614 ms 39.933 ms]
                        change: [-93.187% -93.106% -93.026%] (p = 0.00 < 0.05)
                        Performance has improved.
update rlp node level/size 100000 | depth 5
                        time:   [48.414 ms 50.221 ms 52.693 ms]
                        change: [-94.903% -94.700% -94.444%] (p = 0.00 < 0.05)
                        Performance has improved.
```